add automatic bootstraping of cluster
This commit adds functionality to autobootstrap a percona cluster. It heavily relies on facts and puppetdb to save state information. It starts the service on on all cluster nodes to create users, databases, ...

The first node on the cluster has to be defined as the "bootstrap node". The bootstrap node asks puppetdb to check whether the cluster has
already been started once. If it has not already been started, it starts the service in bootstrap mode and exports it's identity to the other cluster members via puppetdb.
The other cluster members collect that information and start their services if connectivity to the bootstrap node succeeds. They set a local facts to prevent further automatic starts. In addition they export themselves to puppetdb to lock the bootstrap node from bootstrapping again.
If the wsrep_cluster_size on the master is >= 3, the node restarts it's mysql service to exit bootstrap mode. It sets a fact to prevent further starts.
Cluster setup is now complete.